first version of templates SV TagInfo. not working yet
