+++ DataFormats/BTauReco/src/classes.h
@@ -55,6 +55,7 @@ namespace DataFormats_BTauReco {
     std::vector<reco::SecondaryVertexTagInfo::VertexData>               sv_vdv;
     reco::SecondaryVertexTagInfo::IndexedTrackData                      sv_itd;
     std::vector<reco::SecondaryVertexTagInfo::IndexedTrackData>         sv_itdv;
+    std::vector<reco::TemplatedSecondaryVertexTagInfo<reco::IPTagInfo<edm::RefVector<std::vector<reco::Track>,reco::Track,edm::refhelper::FindUsingAdvance<std::vector<reco::Track>,reco::Track> >,reco::JTATagInfo> > ::IndexedTrackData> 						   svngeng;
     reco::SecondaryVertexTagInfo                                        sv;
     reco::SecondaryVertexTagInfoCollection                              sv_c;
     reco::SecondaryVertexTagInfoRef                                     sv_r;
